Store user metadata at object creation
Currently, Carrierwave uses fog's create method to store a new object
to Openstack, an approach that makes it not possible to set user
metadata for a newly created object.

This patch, modifies the above approach to allow setting metadata at
object creation time. To do so, we first create the new object, modify
its metadata using the `fog_attributes' attribute and then save the
object to Openstack with the right headers.