Remove original filename and invalidate memoized filename after save
Carrierwave instructs for memoization in overriden `filename`
methods
in order to support timestamps or random strings in filenames.

However, it expects that memoization will still be present even
after model save and so it does not clean up after the
`original_filename`
attribute in the mounted uploader. As a result, without memoization
any
subsequent saves on the same model instance would end up updating
the
filename column so that the record no longer points to the stored
file.